extract lib
